# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  polling-system_auth_models.User:
    properties:
      created_at:
        type: string
      display_name:
        type: string
      email:
        type: string
      firebase_uid:
        type: string
      id:
        type: string
      role:
        type: string
      updated_at:
        type: string
    type: object
...
  polling-system_polling_models.Answer:
    properties:
      answer:
        type: string
      candidate:
        $ref: '#/definitions/polling-system_polling_models.Candidate'
      candidate_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      question:
        $ref: '#/definitions/polling-system_polling_models.PollQuestion'
      question_id:
        type: string
      source:
        type: string
    type: object
...
  polling-system_polling_models.Candidate:
    properties:
      answers:
        items:
          $ref: '#/definitions/polling-system_polling_models.Answer'
        type: array
      created_at:
        type: string
      id:
        type: string
      phone:
        type: string
      poll:
        $ref: '#/definitions/polling-system_polling_models.Poll'
      poll_id:
        type: string
      updated_at:
        type: string
      voice_poll_sentiment:
        type: string
      voice_poll_started_at:
        type: string
      voicemail_detected:
        type: boolean
    type: object
...
  polling-system_polling_models.ErrorResponse:
    properties:
      error:
        example: something went wrong
        type: string
    type: object
...
  polling-system_polling_models.PaginatedPolls:
    properties:
      items:
        items:
          $ref: '#/definitions/polling-system_polling_models.Poll'
        type: array
      page:
        type: integer
      pages:
        type: integer
      size:
        type: integer
      total:
        type: integer
    type: object
  polling-system_polling_models.PaginatedQuestions:
    properties:
      items:
        items:
          $ref: '#/definitions/polling-system_polling_models.PollQuestion'
        type: array
      page:
        type: integer
      pages:
        type: integer
      size:
        type: integer
      total:
        type: integer
    type: object
  polling-system_polling_models.Poll:
    properties:
      bland_pathway_ids:
        type: object
      candidates:
        items:
          $ref: '#/definitions/polling-system_polling_models.Candidate'
        type: array
      created_at:
        type: string
      created_by:
        type: string
      creator:
        $ref: '#/definitions/polling-system_auth_models.User'
      custom_voice_poll_instructions:
        type: string
      dont_send_sms:
        type: boolean
      first_typeform_sms_text:
        type: string
      id:
        type: string
      only_day_hours:
        type: boolean
      poll_name:
        type: string
      poller_name:
        type: string
      questions:
        items:
          $ref: '#/definitions/polling-system_polling_models.PollQuestion'
        type: array
      scheduled_to_start_at:
        type: string
      status:
        type: string
      time_zone:
        type: string
      typeform_forms:
        type: object
      updated_at:
        type: string
      voice_poll_delay_minutes:
        type: integer
      voice_poll_method:
        type: string
    type: object
  polling-system_polling_models.PollCreate:
    properties:
      bland_pathway_ids:
        type: object
      custom_voice_poll_instructions:
        type: string
      dont_send_sms:
        type: boolean
      first_typeform_sms_text:
        type: string
      only_day_hours:
        type: boolean
      poll_name:
        type: string
      poller_name:
        type: string
      scheduled_to_start_at:
        type: string
      status:
        type: string
      time_zone:
        type: string
      typeform_forms:
        type: object
      voice_poll_delay_minutes:
        type: integer
      voice_poll_method:
        type: string
    required:
    - poll_name
    type: object
  polling-system_polling_models.PollQuestion:
    properties:
      answer_options:
        type: object
      answer_type:
        type: string
      answers:
        items:
          $ref: '#/definitions/polling-system_polling_models.Answer'
        type: array
      created_at:
        type: string
      default_next_question:
        $ref: '#/definitions/polling-system_polling_models.PollQuestion'
      default_next_question_id:
        type: string
      flow_rules:
        type: object
      id:
        type: string
      poll:
        $ref: '#/definitions/polling-system_polling_models.Poll'
      poll_id:
        type: string
      question:
        type: string
      randomize_answers:
        type: boolean
      typeform_field_id:
        type: string
      updated_at:
        type: string
    type: object
  polling-system_polling_models.PollQuestionCreate:
    properties:
      answer_options:
        type: object
      answer_type:
        type: string
      default_next_question_id:
        type: string
      flow_rules:
        type: object
      question:
        type: string
      randomize_answers:
        type: boolean
      typeform_field_id:
        type: string
    required:
    - answer_type
    - question
    type: object
  polling-system_polling_models.PollQuestionUpdate:
    properties:
      answer_options:
        type: object
      answer_type:
        type: string
      default_next_question_id:
        type: string
      flow_rules:
        type: object
      question:
        type: string
      randomize_answers:
        type: boolean
      typeform_field_id:
        type: string
    type: object
  polling-system_polling_models.PollUpdate:
    properties:
      bland_pathway_ids:
        type: object
      custom_voice_poll_instructions:
        type: string
      dont_send_sms:
        type: boolean
      first_typeform_sms_text:
        type: string
      only_day_hours:
        type: boolean
      poll_name:
        type: string
      poller_name:
        type: string
      scheduled_to_start_at:
        type: string
      status:
        type: string
      time_zone:
        type: string
      typeform_forms:
        type: object
      voice_poll_delay_minutes:
        type: integer
      voice_poll_method:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Unified API for authentication, polls, contacts, and file management.
  title: Polling System API
  version: "1.0"
paths:
...
  /polls:
    get:
      description: Returns a paginated list of polls, optionally filtered by status.
      parameters:
      - default: 1
        description: 'Page number (min: 1)'
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page (1-100)
        in: query
        name: size
        type: integer
      - description: Filter by status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/polling-system_polling_models.PaginatedPolls'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/polling-system_polling_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List polls
      tags:
      - Polls
    post:
      consumes:
      - application/json
      description: Creates a new poll. Sets created_by from the authenticated user.
      parameters:
      - description: Poll to create
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/polling-system_polling_models.PollCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/polling-system_polling_models.Poll'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/polling-system_polling_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/polling-system_polling_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a poll
      tags:
      - Polls
  /polls/{id}:
    delete:
      description: Deletes a poll and all its candidates, questions, and answers (cascade).
        Admin only.
      parameters:
      - description: Poll UUID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/polling-system_polling_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a poll
      tags:
      - Polls
    get:
      description: Returns a single poll with its creator.
      parameters:
      - description: Poll UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/polling-system_polling_models.Poll'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/polling-system_polling_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get poll by ID
      tags:
      - Polls
    patch:
      consumes:
      - application/json
      description: Partially updates a poll. Admins can update any poll; pollers can
        only update their own.
      parameters:
      - description: Poll UUID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/polling-system_polling_models.PollUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/polling-system_polling_models.Poll'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/polling-system_polling_models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/polling-system_polling_models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/polling-system_polling_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a poll
      tags:
      - Polls
...
  /polls/{pollId}/questions:
    get:
      description: Returns a paginated list of questions for the given poll.
      parameters:
      - description: Poll UUID
        in: path
        name: pollId
        required: true
        type: string
      - default: 1
        description: 'Page number (min: 1)'
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page (1-100)
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/polling-system_polling_models.PaginatedQuestions'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/polling-system_polling_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List questions for a poll
      tags:
      - Questions
    post:
      consumes:
      - application/json
      description: Adds a question to a poll. Requires poll ownership for pollers.
      parameters:
      - description: Poll UUID
        in: path
        name: pollId
        required: true
        type: string
      - description: Question to add
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/polling-system_polling_models.PollQuestionCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/polling-system_polling_models.PollQuestion'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/polling-system_polling_models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/polling-system_polling_models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/polling-system_polling_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Add question to poll
      tags:
      - Questions
  /questions/{id}:
    delete:
      description: Deletes a question and its answers. Requires ownership of the parent
        poll for pollers.
      parameters:
      - description: Question UUID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/polling-system_polling_models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/polling-system_polling_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a question
      tags:
      - Questions
    get:
      description: Returns a single question.
      parameters:
      - description: Question UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/polling-system_polling_models.PollQuestion'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/polling-system_polling_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get question by ID
      tags:
      - Questions
    patch:
      consumes:
      - application/json
      description: Partially updates a question. Requires ownership of the parent
        poll for pollers.
      parameters:
      - description: Question UUID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/polling-system_polling_models.PollQuestionUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/polling-system_polling_models.PollQuestion'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/polling-system_polling_models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/polling-system_polling_models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/polling-system_polling_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a question
      tags:
      - Questions
...
securityDefinitions:
  BearerAuth:
    description: 'Enter your token in the format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"